typescriptize SavePageControls -wip-
